Fix special price dates, add url_key.
Special price dates are just special_from_date and special_to_date,
we used to collect special_price_*_date which would always be null.
We already created the special_price* fields in Zaius, so we map.

The url_key field is extremely helpful for cart/browse abandon, adding
it to the set of default fields pulled here so we can always capture it.